continuous flow model few wont work

+++ src/main/java/model/Result.java
@@ -5,6 +5,6 @@ import java.util.List;
 public class Result {
     public record ScheduleIntegerResult(List<Integer> start, List<Integer> finish) {
     }
-    public record ScheduleDoubleResult(List<Integer> start, List<Integer> finish) {
+    public record ScheduleDoubleResult(List<Double> start, List<Double> finish) {
     }
 }
